src/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
redis==5.0.0
mysql-connector-python==8.1.0
pymongo==4.5.0
requests==2.31.0
python-dotenv==1.0.0
